requirements.txt
Django>=2.2.5
graphql_core>=2.1.0
graphene>=2.1.5
graphene_django>=2.2.0
django_graphql_jwt>=0.2.1
pytz>=2018.9
graphene-file-upload>=1.2.2
django-filter>=2.1.0
django-filters>=0.2.1
django-cors-headers>=3.0.2
graphql-relay>=2.0.0
Pillow>=5.4.1
djongo>=1.2.36
pymongo>=3.2.0
dataclasses>=0.1
sqlparse==0.2.4